move determine_number_of_jobs into spack.util.cpus, use it in concretize
