+++ var/spack/repos/builtin/packages/gmake/package.py
@@ -6,7 +6,7 @@
 import os
 import re
 
-from spack.build_environment import MakeExecutable, determine_number_of_jobs
+from spack.build_environment import MakeExecutable
 from spack.package import *
